refactor: remove unused code and clean up codebase
Remove unused files, components, and functions to improve maintainability and reduce codebase size.

## Removed Files (16 files, ~1800+ lines)

### Components (7 files)
- BacklogList.tsx - unused backlog display component
- DataMigration.tsx - unused migration tool component
- ImportThoughtsModal.tsx - unused import modal
- PageTransition.tsx - unused animation component
- StaleDataWarning.tsx - unused warning banner
- SummaryPanel.tsx - unused summary dashboard
- SyncStatusIndicator.tsx - unused sync status component

### Hooks (2 files)
- useToolData.ts - unused custom hook
- useVisibilityRefresh.ts - unused visibility handling hook

### Library Files (4 files)
- csvUtils.ts - unused CSV utility functions
- exportRegistry.ts - unused export registry system
- transactionParser.ts - unused transaction parser
- migrations/migrateTaskThoughtIds.ts - unused migration script

### Store Files (2 files)
- useTasksV2.ts - incomplete store refactor
- instances.ts - unused DI pattern implementation

### Types (1 file)
- transactions.ts - moved TransactionCategory to useSpending.ts

### Library Directories
- exportSources/ - entire directory unused (3 files)

### Test Files (3 files)
- TaskStoreIntegration.test.ts - test for deleted useTasksV2
- exportRegistry.test.ts - test for deleted exportRegistry
- transactionParser.test.ts - test for deleted transactionParser

## Modified Files

### deviceDetection.ts
- Removed 7 unused functions (isCapacitorApp, detectDeviceType, detectPlatform, etc.)
- Kept only used functions: parseSource, getDeviceIcon, getDeviceColorClass
- Removed unused types: BrowserType, PlatformType, DeviceInfo interface

### useSpending.ts
- Moved TransactionCategory type definition from deleted transactions.ts
- Type now defined inline where it's actually used

## Verification
- All tests passing (810 tests)
- ESLint check passed
- TypeScript type check passed
