--- src/components/PageTransition.tsx
@@ -1,44 +1,0 @@
-'use client'
-
-import { motion } from 'framer-motion'
-
-interface PageTransitionProps {
-  children: React.ReactNode
-}
-
-const pageVariants = {
-  initial: {
-    opacity: 0,
-    y: 20,
-  },
-  in: {
-    opacity: 1,
-    y: 0,
-  },
-  out: {
-    opacity: 0,
-    y: -20,
-  },
-}
-
-const pageTransition = {
-  type: 'tween',
-  ease: 'anticipate',
-  duration: 0.4,
-}
-
-export function PageTransition({ children }: PageTransitionProps) {
-  return (
-    <motion.div
-      initial="initial"
-      animate="in"
-      exit="out"
-      variants={pageVariants}
-      transition={pageTransition}
-      className="w-full"
-    >
-      {children}
-    </motion.div>
-  )
-}
-
